Revert "Refactor ext-proc Main with Server Package Add Hermetic Test with k8s…"
This reverts commit d7fdc8fe10c810a96a00ebfa884a36e6ef9bba5f.
